requirements.txt
# Core dependencies
pinecone-client>=6.0.0
langchain>=0.3.25
langchain-community>=0.3.24
python-dotenv>=1.1.0

# Document processing
pypdf>=5.5.0
sentence-transformers>=4.1.0

# Google Cloud
google-cloud-storage>=3.1.0
google-auth>=2.40.2
